+++ internal/storage/issueops/journal_completeness_test.go
@@ -75,6 +75,7 @@ var mutationEntryPoints = []string{
 	// update
 	"UpdateIssueInTx",
 	"UpdateIssueWithoutEventInTx",
+	"TouchIssueInTx",
 	"MergeMetadataInTx",
 	"DeleteMetadataInTx",
 	"ApplyLabelPatch",